pyproject.toml
[project]
name = "writeup-chatbot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.119.0",
    "jinja2>=3.1.6",
    "openai>=2.3.0",
    "pydantic>=2.12.2",
    "python-dotenv>=1.1.1",
    "uvicorn>=0.37.0",
]
